// frontend/src/components/rich-editor/types.ts
export interface FormattingState {
  /** Whether bold is active */
  isBold: boolean;
  /** Whether bold command can be executed */
  canBold: boolean;
  /** Whether italic is active */
  isItalic: boolean;
  /** Whether italic command can be executed */
  canItalic: boolean;
  /** Whether underline is active */
  isUnderline: boolean;
  /** Whether underline command can be executed */
  canUnderline: boolean;
  /** Whether strike is active */
  isStrike: boolean;
  /** Whether strike command can be executed */
  canStrike: boolean;
  /** Whether code is active */
  isCode: boolean;
  /** Whether code command can be executed */
  canCode: boolean;
  /** Whether marks can be cleared */
  canClearMarks: boolean;
  /** Whether paragraph is active */
  isParagraph: boolean;
  /** Whether heading level 1 is active */
  isHeading1: boolean;
  /** Whether heading level 2 is active */
  isHeading2: boolean;
  /** Whether heading level 3 is active */
  isHeading3: boolean;
  /** Whether heading level 4 is active */
  isHeading4: boolean;
  /** Whether heading level 5 is active */
  isHeading5: boolean;
  /** Whether heading level 6 is active */
  isHeading6: boolean;
  /** Whether bullet list is active */
  isBulletList: boolean;
  /** Whether ordered list is active */
  isOrderedList: boolean;
  /** Whether code block is active */
  isCodeBlock: boolean;
  /** Whether blockquote is active */
  isBlockquote: boolean;
  /** Whether link is active */
  isLink: boolean;
  /** Whether link command can be executed */
  canLink: boolean;
  /** Whether undo can be executed */
  canUndo: boolean;
  /** Whether redo can be executed */
  canRedo: boolean;
  /** Whether table is active */
  isTable: boolean;
  /** Whether table command can be executed */
  canTable: boolean;
  /** Whether image command can be executed */
  canImage: boolean;
  /** Whether math is active */
  isMath: boolean;
  /** Whether inline math is active */
  isInlineMath: boolean;
  /** Whether math command can be executed */
  canMath: boolean;
  /** Whether taskList is active */
  isTaskList: boolean;
}